# app/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Identity Atlas Ingest API
  version: 3.4.0
  description: |
    REST API for ingesting authorization data from any source system into Identity Atlas.
    Crawlers authenticate with API keys (Bearer token) and POST batches of records.
    The API handles validation, bulk merge, scoped delete detection, and temporal versioning.

    **Scope:** this specification covers only the **ingest + crawler-admin** surface —
    the endpoints crawlers and the crawler-management UI call. The authenticated
    **read API** consumed by the web UI (matrix & permissions, identities, resources,
    governance & business roles, risk scores, contexts, entity detail, org-chart, tags,
    account-linking — 100+ routes) is **not** described here; see the API reference under
    `docs/api/` for those.
  contact:
    name: Fortigi
    url: https://github.com/Fortigi/IdentityAtlas
servers:
  - url: /api
    description: Relative to deployment root
security:
  - CrawlerApiKey: []
...
paths:
...
  /nl-reports/catalog:
    get:
      tags: [Custom Reports]
      summary: Report builder catalog
      operationId: getCustomReportCatalog
      description: >
        The entities, fields, relations, operators and pickable columns a report definition may
        use, with the known values of enumerated fields. Alongside the fields every deployment has,
        this includes the `extendedAttributes` keys discovered in this deployment's own data, named
        `ext.<rawKey>` and marked `discovered`. Requires `data.write.reports`; answers 404 while the
        `customReports` feature is off.

      security:
        - EntraIdBearer: []
      responses:
        '200':
          description: Builder catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  entities:
                    type: object
                    description: Keyed by entity name.
                    additionalProperties:
                      type: object
                      properties:
                        label: { type: string }
                        table: { type: string }
                        description: { type: string }
                        compareRelations:
                          type: array
                          items: { type: string }
                        defaultColumns:
                          type: array
                          items: { type: string }
                        fields:
                          type: array
                          items:
                            type: object
                            properties:
                              name: { type: string }
                              label: { type: string }
                              type: { type: string }
                              values:
                                type: array
                                description: Known values, for enumerated fields only.
                                items: { type: string }
                              discovered:
                                type: boolean
                                description: >
                                  Present on a field read from this deployment's own
                                  `extendedAttributes` rather than from the shared catalog.
                        relations:

                          type: array
                          items:
                            type: object
                            properties:
                              name: { type: string }
                              label: { type: string }
                              target: { type: string }
                              cardinality: { type: string }
                        columns:
                          type: array
                          items:
                            $ref: '#/components/schemas/ReportColumn'
                        groupableFields:
                          type: array
                          description: The fields this entity can be counted per (no date fields).
                          items:
                            type: object
                            properties:
                              name: { type: string }
                              label: { type: string }
                              discovered: { type: boolean }
                  operators:

                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        label: { type: string }
                        needsValue: { type: boolean }
                  operatorsByType:
                    type: object
                    description: Operator names allowed per field type.
                    additionalProperties:
                      type: array
                      items: { type: string }
                  compareMeasures:
                    type: object
                    description: Compare measure name → label.
                    additionalProperties: { type: string }
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '500':
          $ref: '#/components/responses/CustomReportsFailed'
  /nl-reports/lookup:
    get:
      tags: [Custom Reports]
      summary: Search object names for a compare reference
      operationId: lookupCustomReportNames
      description: >
        Fuzzy name search used by the builder's reference picker. A search text shorter than 2
        or longer than 100 characters returns an empty list. Requires `data.write.reports`;
        answers 404 while the `customReports` feature is off.
      security:
        - EntraIdBearer: []
      parameters:
        - name: entity
          in: query
          required: true
          description: Entity name from the catalog.
          schema: { type: string }
        - name: q
          in: query
          required: true
          description: Search text.
          schema: { type: string }
      responses:
        '200':
          description: Best matches, most similar first (at most 5)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        type:
                          type: string
                          nullable: true
                        score: { type: number }
        '400':
          description: Unknown entity
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '500':
          $ref: '#/components/responses/CustomReportsFailed'
  /nl-reports/status:
    get:
      tags: [Custom Reports]
      summary: Report generator availability
      operationId: getCustomReportGeneratorStatus
      description: >
        Whether the local model server is reachable and serves the configured model. Never fails
        on an unreachable server — it reports `available: false` with a reason instead. Requires
        `data.write.reports`; answers 404 while the `customReports` feature is off.
      security:
        - EntraIdBearer: []
      responses:
        '200':
          description: Generator status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available: { type: boolean }
                  model:
                    type: string
                    nullable: true
                  loaded: { type: boolean }
                  promptCache:
                    type: string
                    enum: [cold, warming, ready, failed]
                  reason:
                    type: string
                    nullable: true
                    description: "`model-not-installed` or `server-unreachable`; null when available."
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
  /nl-reports/warm:
    post:
      tags: [Custom Reports]
      summary: Warm up the report generator
      operationId: warmCustomReportGenerator
      description: >
        Loads the configured model and restores its prompt cache. Waits at most a few seconds:
        when the warm-up is still running it answers `state: preparing` and continues in the
        background. Requires `data.write.reports`; answers 404 while the `customReports` feature
        is off.
      security:
        - EntraIdBearer: []
      responses:
        '200':
          description: Warm-up finished (`ready`) or still running (`preparing`)
          content:
            application/json:
              schema:
                type: object
                required: [state]
                properties:
                  state:
                    type: string
                    enum: [ready, preparing]
                  model:
                    type: string
                    description: Present when ready.
                  ms:
                    type: integer
                    description: Warm-up duration, present when ready.
                  restored:
                    type: boolean
                    description: Whether a saved prompt cache was restored (llama.cpp backend only).
                  message:
                    type: string
                    description: Present while preparing.
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '502':
          $ref: '#/components/responses/ModelServerFailed'
  /nl-reports/interpret:
    post:
      tags: [Custom Reports]
      summary: Turn a question into a report definition
      operationId: interpretCustomReportQuestion
      description: >
        Sends a plain-language question (plus earlier conversation turns) to the local model and
        returns a validated report definition (`report`), a definition whose named objects need
        confirming (`confirm`), a clarifying question (`clarify`), or `error` when the model's
        answer could not be used. Requires `data.write.reports`; answers 404 while the
        `customReports` feature is off.
      security:
        - EntraIdBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  maxLength: 2000
                history:
                  type: array
                  maxItems: 12
                  description: Earlier turns, at most 10,000 characters of content in total.
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                        maxLength: 20000
                model:
                  type: string
                  pattern: '^[A-Za-z0-9._:/-]{1,100}$'
                  description: Evaluation override of the configured model; the UI never sends it.
      responses:
        '200':
          description: The model's interpretation
          content:
            application/json:
              schema:
                type: object
                required: [kind]
                properties:
                  kind:
                    type: string
                    enum: [report, confirm, clarify, error]
                  spec:
                    $ref: '#/components/schemas/ReportDefinition'
                  confirm:
                    $ref: '#/components/schemas/ReportConfirmation'
                  explanation:
                    $ref: '#/components/schemas/ReportExplanation'
                  assumptions:
                    type: array
                    items: { type: string }
                  warnings:
                    type: array
                    items: { type: string }
                  sql:
                    type: string
                    description: Compiled SQL (`report` only).
                  question:
                    type: string
                    description: Clarifying question (`clarify` only).
                  options:
                    type: array
                    description: Up to 4 suggested answers (`clarify` only).
                    items: { type: string }
                  message:
                    type: string
                    description: What went wrong (`error` only).
                  errors:
                    type: array
                    items: { type: string }
                  raw:
                    type: string
                    description: The model's raw reply.
                  timing:
                    type: object
                    additionalProperties: true
                  model: { type: string }
                  repaired:
                    type: boolean
                    description: Whether a repair round was needed.
        '400':
          $ref: '#/components/responses/CustomReportsBadRequest'
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '429':
          description: The caller's previous question is still being answered (one at a time per user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '502':
          $ref: '#/components/responses/ModelServerFailed'
  /nl-reports/resolve:
    post:
      tags: [Custom Reports]
      summary: Apply a confirmation answer and look named objects up again
      operationId: resolveCustomReportReferences
      description: >
        Validates a report definition, applies the analyst's answer to a "did you mean"
        confirmation (when `choice` is given), and looks up the objects it names again. No model
        is involved. Requires `data.write.reports`; answers 404 while the `customReports` feature
        is off.
      security:
        - EntraIdBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec:
                  $ref: '#/components/schemas/ReportDefinition'
                choice:
                  type: object
                  required: [path, name]
                  properties:
                    path:
                      type: array
                      description: The `path` of the confirmation being answered.
                      items: { type: integer }
                    name:
                      type: string
                      description: The chosen or typed exact name.
                    id:
                      type: string
                      description: The chosen record, when known.
                    keep:
                      type: boolean
                      description: Use a "name is X" value as written, without checking it again.
      responses:
        '200':
          description: The updated definition and the next confirmation, if any
          content:
            application/json:
              schema:
                type: object
                properties:
                  spec:
                    $ref: '#/components/schemas/ReportDefinition'
                  confirm:
                    allOf:
                      - $ref: '#/components/schemas/ReportConfirmation'
                    nullable: true
                  explanation:
                    $ref: '#/components/schemas/ReportExplanation'
        '400':
          $ref: '#/components/responses/CustomReportsBadRequest'
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '500':
          $ref: '#/components/responses/CustomReportsFailed'
  /nl-reports/run:
    post:
      tags: [Custom Reports]
      summary: Run a report definition
      operationId: runCustomReportDefinition
      description: >
        Validates and runs a report definition in a read-only transaction with a statement
        timeout. A definition that is invalid, or names an object that still needs confirming,
        answers 400 with the errors, the confirmation and the validated definition. Requires
        `data.write.reports`; answers 404 while the `customReports` feature is off.
      security:
        - EntraIdBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec:
                  $ref: '#/components/schemas/ReportDefinition'
      responses:
        '200':
          description: Report rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  spec:
                    $ref: '#/components/schemas/ReportDefinition'
                  explanation:
                    $ref: '#/components/schemas/ReportExplanation'
                  columns:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportColumn'
                  rows:
                    type: array
                    description: >
                      One object per row, keyed by column key, with an `_entity` property
                      ({kind, id}) linking the row to an entity detail page.
                    items:
                      type: object
                      additionalProperties: true
                  total:
                    type: integer
                  truncated:
                    type: boolean
                    description: More rows matched than the definition's limit.
                  sql:
                    type: string
                  params:
                    type: array
                    items: {}
                  elapsedMs:
                    type: integer
        '400':
          description: Missing or invalid definition, or a named object needs confirming
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  errors:
                    type: array
                    items: { type: string }
                  confirm:
                    $ref: '#/components/schemas/ReportConfirmation'
                  spec:
                    $ref: '#/components/schemas/ReportDefinition'
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '500':
          $ref: '#/components/responses/CustomReportsFailed'
  /nl-reports/saved:
    post:
      tags: [Custom Reports]
      summary: Save a new custom report
      operationId: createSavedCustomReport
      description: >
        Stores a named report definition. Saved reports are listed, run and downloaded through
        `/reports` as `custom-<id>`. Requires `data.write.reports`; answers 404 while the
        `customReports` feature is off.
      security:
        - EntraIdBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedReportInput'
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedReport'
        '400':
          $ref: '#/components/responses/CustomReportsBadRequest'
        '403':
          $ref: '#/components/responses/CustomReportsForbidden'
        '404':
          $ref: '#/components/responses/CustomReportsDisabled'
        '409':
          description: A report with that name already exists
        '500':
          $ref: '#/components/responses/CustomReportsFailed'
...
components:
  securitySchemes:
    CrawlerApiKey:
      type: http
      scheme: bearer
      description: "Crawler API key (format: fgc_<random>)"
    EntraIdBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entra ID JWT token
  schemas:
...
    ReportColumn:
      type: object
      required: [key, label]
      properties:
        key:
          type: string
        label:
          type: string
        discovered:
          type: boolean
          description: >
            Only in the builder catalog: the column reads an `extendedAttributes` key found in this
            deployment's own data rather than a field of the shared catalog.
...
    ReportConfirmation:
      type: object
      description: >
        A named object in the definition that did not match exactly — the analyst picks one of
        `choices` or types the exact name, and sends it to `/nl-reports/resolve`.
      properties:
        kind:
          type: string
          enum: [reference, value]
        path:
          type: array
          description: Index path to the condition in the definition.
          items:
            type: integer
        name:
          type: string
        label:
          type: string
        message:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              type:
                type: string
                nullable: true
              score:
                type: number
...
  responses:
...
    CustomReportsForbidden:
      description: The caller lacks the required permission (checked before the feature flag)
    CustomReportsDisabled:
      description: The `customReports` feature is not enabled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
    CustomReportsFailed:
      description: The request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    ModelServerFailed:
      description: The local model server is not reachable or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'